requirements.txt
pandas
numpy
gradio
langchain
langchain-community
langchain-openai
langchain-text-splitters
chromadb
python-dotenv
